# api-contracts/openapi/agent_api.yaml
openapi: 3.0.3
info:
  title: SMT Python Agent REST API
  description: |
    SMT 智能运维平台 Python 智能体对外 RESTful 接口契约。

    覆盖两个 Python Agent：
    - agent-knowledge（端口 8004）：知识助手 RAG 问答，支持文档上传、向量检索问答、文档列表与删除。
    - agent-maintenance（端口 8002）：设备运维 Agent，支持健康评分、故障诊断、趋势预测、案例沉淀。

    外部访问路径前缀 `/api/agent/**` 经 `smt-gateway` 网关路由到对应 Python 服务，
    网关 `StripPrefix=2` 后 Python 服务收到 `/knowledge/**` 与 `/maintenance/**`。
    响应体为业务对象直返（不使用 Result 包装），字段命名采用 snake_case 与 Python Pydantic 默认对齐。
  version: 1.0.0
  contact:
    name: smt-agent-platform
    url: https://github.com/north30/smt-agent-platform
  license:
    name: Apache-2.0
servers:
  - url: http://localhost:8080
    description: 本地开发环境（经 smt-gateway 网关）
...
paths:
...
  /api/agent/quality/monitor/{device_id}:
    get:
      tags:
        - 质量分析
      summary: 不良实时监控
      description: |
        拉取设备最新 AOI/SPI 检测类数据点，按阈值实时计算不良率与告警状态。
        超阈值告警同步落库到 PostgreSQL `quality_alerts` 表。
        依赖 smt-device-service 提供数据点，Java 服务不可达时返回 503；
        当设备近 1 小时无 AOI 数据点时返回 `status=INSUFFICIENT_DATA`。
      operationId: getQualityMonitor
      parameters:
        - $ref: '#/components/parameters/DeviceIdPathParam'
      responses:
        '200':
          description: 监控数据返回成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QualityMonitorResponse'
              examples:
                alert:
                  summary: 超阈值告警示例
                  value:
                    device_id: 1001
                    defect_rate: 0.025
                    threshold: 0.02
                    status: ALERT
                    datapoints:
                      - datapoint_code: AOI_DEFECT_RATE
                        value: 0.025
                        timestamp: '2026-07-06T08:30:00Z'
                    analyzed_at: '2026-07-06T08:30:05Z'
        '503':
          description: 依赖的 Java 服务（smt-device-service）不可达
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                upstream:
                  summary: Java 服务不可达
                  value:
                    error: DEVICE_SERVICE_UNAVAILABLE
                    message: 设备服务暂不可用
  /api/agent/quality/root_cause:
    post:
      tags:
        - 质量分析
      summary: 根因分析
      description: |
        基于"人/机/料/法/环"五要素框架，结合设备信息 + 历史质量案例 RAG + 大模型推理，
        输出根因假设列表与纠正措施。空案例库时仍返回 LLM 推理结果，similar_cases 返回空列表。
      operationId: analyzeQualityRootCause
      requestBody:
        required: true
        description: 根因分析请求体
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QualityRootCauseRequest'
            examples:
              offset:
                summary: 贴片偏移根因分析
                value:
                  device_id: 1001
                  defect_description: 贴片偏移率 5%
      responses:
        '200':
          description: 根因分析成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QualityRootCauseResponse'
              examples:
                rootCause:
                  summary: 根因分析结果示例
                  value:
                    device_id: 1001
                    root_causes:
                      - category: 机
                        cause: 贴装头 Z 轴校准漂移
                      - category: 料
                        cause: 锡膏厚度异常
                    corrective_actions:
                      - 重新校准贴装头 Z 轴
                      - 检查钢网开口与锡膏黏度
                    similar_cases:
                      - case_id: qc-0023
                        description: 贴片偏移伴随锡膏厚度异常
                        root_cause: 钢网开口偏移 + 锡膏搅拌不均
                        corrective_action: 重新制作钢网并规范锡膏搅拌流程
                        score: 0.85
        '400':
          description: 参数错误（device_id 缺失或 defect_description 为空）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                emptyDescription:
                  summary: 缺陷描述为空
                  value:
                    error: INVALID_DEFECT_DESCRIPTION
                    message: 缺陷描述不能为空
        '503':
          description: 依赖的 LLM 服务或向量库不可达
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                upstream:
                  summary: 上游服务不可达
                  value:
                    error: LLM_UNAVAILABLE
                    message: 上游服务暂不可用
  /api/agent/quality/cases:
    post:
      tags:
        - 质量分析
      summary: 录入质量案例
      description: |
        将历史质量案例（缺陷类型、描述、根因、纠正措施）向量化入库到 Milvus `smt_quality_cases` collection，
        供后续根因分析 RAG 检索复用。返回新案例 ID。
      operationId: createQualityCase
      requestBody:
        required: true
        description: 质量案例创建请求体
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QualityCaseCreateRequest'
            examples:
              case:
                summary: 案例创建示例
                value:
                  defect_type: OFFSET
                  description: 贴片偏移率 5%，主要发生在 PCB 右下角
                  root_cause: 钢网开口偏移 + 锡膏搅拌不均
                  corrective_action: 重新制作钢网并规范锡膏搅拌流程
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QualityCaseCreateResponse'
              examples:
                success:
                  summary: 创建成功
                  value:
                    case_id: qc-0024
        '400':
          description: 参数错误（必填字段缺失）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missingField:
                  summary: 必填字段缺失
                  value:
                    error: MISSING_CASE_REQUIRED_FIELDS
                    message: 缺陷类型、描述、根因、纠正措施均为必填
  /api/agent/quality/alerts:
    get:
      tags:
        - 质量分析
      summary: 查询质量告警
      description: 分页查询历史质量告警记录（来自 PostgreSQL `quality_alerts` 表），按告警时间倒序。
      operationId: listQualityAlerts
      parameters:
        - name: page
          in: query
          description: 页码（从 1 开始）
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
            example: 1
        - name: size
          in: query
          description: 每页大小
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 20
            example: 20
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QualityAlertsPageResponse'
              examples:
                page:
                  summary: 告警分页示例
                  value:
                    records:
                      - id: 158
                        device_id: 1001
                        defect_rate: 0.025
                        threshold: 0.02
                        status: ALERT
                        datapoint_code: AOI_DEFECT_RATE
                        alert_time: '2026-07-06T08:30:05Z'
                    total: 1
                    page: 1
                    size: 20
        '400':
          description: 参数错误（page/size 不合法）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidPage:
                  summary: 分页参数不合法
                  value:
                    error: INVALID_PAGINATION
                    message: page 须 ≥1，size 须在 1-100 之间
...
components:
  parameters:
...
    DeviceIdPathParam:
      name: device_id
      in: path
      description: 设备 ID
      required: true
      schema:
        type: integer
        format: int64
        example: 1001
...
  schemas:
    ErrorResponse:
      type: object
      description: 错误响应
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: 错误码（大写蛇形，如 INVALID_FILE、DEVICE_NOT_FOUND）
          example: INVALID_FILE
        message:
          type: string
          description: 错误描述
          example: 上传文件不能为空
...
    QualityDatapoint:
      type: object
      description: 质量采集点数据
      required:
        - datapoint_code
        - value
        - timestamp
      properties:
        datapoint_code:
          type: string
          description: 采集点编码（如 AOI_DEFECT_RATE、SPI_SOLDER_PASTE_VOLUME）
          example: AOI_DEFECT_RATE
        value:
          type: number
          format: float
          description: 采集值
          example: 0.025
        timestamp:
          type: string
          format: date-time
          description: 采集时间（ISO-8601）
          example: '2026-07-06T08:30:00Z'
    QualityMonitorResponse:
      type: object
      description: 不良实时监控响应
      required:
        - device_id
        - defect_rate
        - threshold
        - status
        - datapoints
        - analyzed_at
      properties:
        device_id:
          type: integer
          format: int64
          description: 设备 ID
          example: 1001
        defect_rate:
          type: number
          format: float
          description: 当前不良率（0-1）
          minimum: 0
          example: 0.025
        threshold:
          type: number
          format: float
          description: 不良率阈值
          minimum: 0
          example: 0.02
        status:
          type: string
          description: 告警状态
          enum:
            - OK
            - ALERT
            - INSUFFICIENT_DATA
          example: ALERT
        datapoints:
          type: array
          description: 命中的采集点数据列表
          items:
            $ref: '#/components/schemas/QualityDatapoint'
        analyzed_at:
          type: string
          format: date-time
          description: 分析时间（ISO-8601）
          example: '2026-07-06T08:30:05Z'
    QualityRootCauseRequest:
      type: object
      description: 根因分析请求体
      required:
        - device_id
        - defect_description
      properties:
        device_id:
          type: integer
          format: int64
          description: 设备 ID
          example: 1001
        defect_description:
          type: string
          description: 缺陷描述
          example: 贴片偏移率 5%
    QualityRootCause:
      type: object
      description: 根因假设（按人/机/料/法/环分类）
      required:
        - category
        - cause
      properties:
        category:
          type: string
          description: 五要素分类（人/机/料/法/环）
          example: 机
        cause:
          type: string
          description: 根因描述
          example: 贴装头 Z 轴校准漂移
    QualitySimilarCase:
      type: object
      description: 相似历史质量案例
      required:
        - case_id
        - description
        - root_cause
        - corrective_action
        - score
      properties:
        case_id:
          type: string
          description: 案例 ID
          example: qc-0023
        description:
          type: string
          description: 案例描述
          example: 贴片偏移伴随锡膏厚度异常
        root_cause:
          type: string
          description: 案例根因
          example: 钢网开口偏移 + 锡膏搅拌不均
        corrective_action:
          type: string
          description: 案例纠正措施
          example: 重新制作钢网并规范锡膏搅拌流程
        score:
          type: number
          format: float
          description: 相似度得分（0-1）
          minimum: 0
          maximum: 1
          example: 0.85
    QualityRootCauseResponse:
      type: object
      description: 根因分析响应
      required:
        - device_id
        - root_causes
        - corrective_actions
        - similar_cases
      properties:
        device_id:
          type: integer
          format: int64
          description: 设备 ID
          example: 1001
        root_causes:
          type: array
          description: 根因假设列表（按五要素分类）
          items:
            $ref: '#/components/schemas/QualityRootCause'
        corrective_actions:
          type: array
          description: 纠正措施列表
          items:
            type: string
          example:
            - 重新校准贴装头 Z 轴
            - 检查钢网开口与锡膏黏度
        similar_cases:
          type: array
          description: 相似历史案例列表
          items:
            $ref: '#/components/schemas/QualitySimilarCase'
    QualityCaseCreateRequest:
      type: object
      description: 质量案例创建请求体
      required:
        - defect_type
        - description
        - root_cause
        - corrective_action
      properties:
        defect_type:
          type: string
          description: 缺陷类型（如 OFFSET、SHORT、OPEN、MISSING）
          example: OFFSET
        description:
          type: string
          description: 缺陷描述
          example: 贴片偏移率 5%，主要发生在 PCB 右下角
        root_cause:
          type: string
          description: 根因
          example: 钢网开口偏移 + 锡膏搅拌不均
        corrective_action:
          type: string
          description: 纠正措施
          example: 重新制作钢网并规范锡膏搅拌流程
    QualityCaseCreateResponse:
      type: object
      description: 质量案例创建响应
      required:
        - case_id
      properties:
        case_id:
          type: string
          description: 新建案例 ID
          example: qc-0024
    QualityAlertRecord:
      type: object
      description: 质量告警记录
      required:
        - id
        - device_id
        - defect_rate
        - threshold
        - status
        - alert_time
      properties:
        id:
          type: integer
          format: int64
          description: 告警记录 ID
          example: 158
        device_id:
          type: integer
          format: int64
          description: 设备 ID
          example: 1001
        defect_rate:
          type: number
          format: float
          description: 告警时不良率
          example: 0.025
        threshold:
          type: number
          format: float
          description: 不良率阈值
          example: 0.02
        status:
          type: string
          description: 告警状态
          example: ALERT
        datapoint_code:
          type: string
          description: 触发告警的采集点编码（无具体采集点时为 null）
          nullable: true
          example: AOI_DEFECT_RATE
        alert_time:
          type: string
          format: date-time
          description: 告警时间（ISO-8601）
          example: '2026-07-06T08:30:05Z'
    QualityAlertsPageResponse:
      type: object
      description: 质量告警分页响应
      required:
        - records
        - total
        - page
        - size
      properties:
        records:
          type: array
          description: 告警记录列表（按告警时间倒序）
          items:
            $ref: '#/components/schemas/QualityAlertRecord'
        total:
          type: integer
          description: 总记录数
          example: 1
        page:
          type: integer
          description: 当前页码
          example: 1
        size:
          type: integer
          description: 每页大小
          example: 20
...